UIMARCAUTH-122 correctly scroll the details pane (#127)
* UIMARCAUTH-122 correctly scroll the details pane

The only direct descendents of a `<Paneset>` should be `<Pane>`
components. Here, an extraneous `<div>` wrapper was inserted for the
purpose of providing a test hook. This change resolves the issue by
hooking directly onto the `test-id` of the inner `<Pane>`. This feels a
little risky since that `test-id` is not directly controlled by this
application.

A more thorough fix might be to allow the test-id in `@folio/quick-marc`
to be overriden by prop.

Refs UIMARCAUTH-122

* be less greedy about pane-width

Co-authored-by: Denys Bohdan <denys_bohdan@epam.com>